Learn to Program with C
Written by Noel Kalicharan. It is an excellent book for beginners with Lessons, Examples, and core concepts. It has clean look and is easy to understand. It has the basics of programming concepts to the concept of the structures. All basic syntax, logic, functions, arrays, searching, sorting, and merging are covered. Thus, a beginner can be highly benefited from this book.